Handle image name resolving for containerd format
The YAML input file expects image names in the standard Docker client
format, and are shared between Docker driver and containerd driver use
cases. This commit adds a function (mostly copied from Docker) that
takes a "bare" image name and decorates it as necessary to be useful for
containerd gRPC calls.

Signed-off-by: Phil Estes <estesp@gmail.com>